requirementes.txt
asgiref==3.6.0
bcc==0.18.0
beautifulsoup4==4.11.1
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
Django==3.2.17
django-allauth==0.52.0
django-bootstrap4==22.3
django-crispy-forms==1.14.0
django-extra-views==0.14.0
django-filter==22.1
django-htmx==1.13.0
django-tables2==2.5.1
heroicons==2.1.1
hidpidaemon==18.4.6
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
more-itertools==8.10.0
netaddr==0.8.0
netifaces==0.11.0
oauthlib==3.2.0
Pillow==9.4.0
pop-transition==1.1.2
protobuf==3.12.4
psycopg2==2.9.5
pycairo==1.20.1
pycodestyle==2.10.0
pycups==2.0.1
pydbus==0.6.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0
python-debian===0.1.43ubuntu1
python-gnupg==0.4.8
python-xlib==0.29
python3-openid==3.2.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
repolib==2.2.0
repoman==1.4.0
requests==2.25.1
requests-oauthlib==1.3.1
SecretStorage==3.3.1
sessioninstaller==0.0.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
systemd-python==234
tomli==2.0.1
ubuntu-advantage-tools==27.12
ubuntu-drivers-common==0.0.0
ufw==0.36.1
urllib3==1.26.5
wadllib==1.3.6
xdg==5
xkit==0.0.0
zipp==1.0.0
